src/_example/fastapi/pyproject.toml
[tool.poetry]
name = "forestadmin-fastapi-demo"
version = "0.1.0-beta.15"
description = ""
authors = ["Julien Barreau <julien.barreau@forestadmin.com>",]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
sqlalchemy = ">=1.4.0"
aiosqlite = ">=0.20.0"
greenlet = ">=3.0.3"
fastapi = ">=0.110.0"
uvicorn = ">=0.11.5"
typer = ">=0.12.1"
# SQLModel =
python-dotenv = ">= 1.0.0"
forestadmin-datasource-toolkit = {path = "../../datasource_toolkit", develop = true}
forestadmin-agent-toolkit = {path = "../../agent_toolkit", develop = true}
forestadmin-datasource-sqlalchemy = {path = "../../datasource_sqlalchemy", develop = true}
# forestadmin-agent-flask = {path = "../../flask_agent", develop = true}
# forestadmin-agent-fastapi = {path = "../../fastapi_agent", develop = true}

[tool.poetry.group.test]
optional = true

[tool.poetry.group.linter]
optional = true

[tool.poetry.group.formatter]
optional = true

[tool.poetry.group.sorter]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "~=7.1"
pytest-asyncio = "~=0.18"
coverage = "~=6.5"
freezegun = "~=1.2.0"

[[tool.poetry.group.test.dependencies.mock]]
version = "4.0"
python = "3.8"

[tool.poetry.group.linter.dependencies]
[[tool.poetry.group.linter.dependencies.flake8]]
version = "~=5.0"
python = "<3.8.1"

[[tool.poetry.group.linter.dependencies.flake8]]
version = "~=6.0"
python = ">=3.8.1"

[tool.poetry.group.formatter.dependencies]
black = ">=24.3.0"

[tool.poetry.group.sorter.dependencies]
isort = "~=3.6"

[tool.poe.executor]
type = "simple"

[tool.poe.tasks]
# init-db  = { script = "demo.models.utils.init_db:create_all" }
# populate-db  = { script = "demo.models.utils.populate_dev_db:populate" }
# runserver = { shell = "FLASK_APP=demo.app flask run -h 0.0.0.0 --debugger --reload" }

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.16.5"
faker = "^15.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
